+++ lua/differ/init.lua
@@ -66,6 +66,7 @@ end
 ---@field extra_keymaps? differ.panel.ExtraMap[]
 ---@field on_rerender? fun()
 ---@field on_cursor? fun()
+---@field on_repurpose? fun(buf: integer): boolean
 
 ---@param model differ.DiffModel
 ---@param opts? differ.DiffModelOpts
@@ -89,6 +90,7 @@ function M.diff_model(model, opts)
             extra_keymaps = opts.extra_keymaps,
             on_rerender = opts.on_rerender,
             on_cursor = opts.on_cursor,
+            on_repurpose = opts.on_repurpose,
         })
         :open()
 end